UMA2 resource API calls into the resource module
