+++ src/uma2/mod.rs
@@ -3,6 +3,8 @@ mod claim_token_format;
 mod permission_association;
 mod provider;
 mod error;
+mod permission_ticket;
+mod resource;
 
 pub use claim_token_format::Uma2ClaimTokenFormat;
 pub use permission_association::{
@@ -12,5 +14,5 @@ pub use permission_association::{
 };
 pub use provider::Uma2Provider;
 pub use error::Uma2Error;
-
-// pub use uma2::
+pub use permission_ticket::Uma2PermissionTicket;
+pub use resource::Uma2Resource;
